LayoutTests/http/tests/security/blob-null-url-location-origin-expected.txt

CONSOLE MESSAGE: data URL frame loaded
CONSOLE MESSAGE: blob iframe opened
CONSOLE MESSAGE: blob iframe loadednull
CONSOLE MESSAGE: blob iframe message posted
PASS
